medical_bot/requirements.txt
pyTelegramBotAPI>=3.0.0
psycopg2-binary
python-dateutil==2.8.2